refactor(QuestionType): Change format of QuestionType Enum
