+++ src/main/scala/com/particeep/api/models/enums/QuestionType.scala
@@ -1,20 +1,18 @@
 package com.particeep.api.models.enums
 
-object QuestionType {
+sealed trait QuestionType extends Product with Enum
 
-  sealed abstract class QuestionType extends Product with Enum
-  case object TEXT                   extends QuestionType
-  case object LONGTEXT               extends QuestionType
-  case object DATE                   extends QuestionType
-  case object RADIO                  extends QuestionType
-  case object SELECT                 extends QuestionType
-  case object CHECKBOX               extends QuestionType
-  case object DOCUMENT               extends QuestionType
-  case object LABEL                  extends QuestionType
+object QuestionType extends EnumHelper[QuestionType] {
+  case object TEXT     extends QuestionType
+  case object LONGTEXT extends QuestionType
+  case object DATE     extends QuestionType
+  case object RADIO    extends QuestionType
+  case object SELECT   extends QuestionType
+  case object CHECKBOX extends QuestionType
+  case object DOCUMENT extends QuestionType
+  case object LABEL    extends QuestionType
 
-  object QuestionType extends EnumHelper[QuestionType] {
-    def values: Set[QuestionType] = Set(TEXT, LONGTEXT, DATE, RADIO, SELECT, CHECKBOX, DOCUMENT, LABEL)
+  def values: Set[QuestionType] = Set(TEXT, LONGTEXT, DATE, RADIO, SELECT, CHECKBOX, DOCUMENT, LABEL)
 
-    implicit def stringToQuestionType(value: String): QuestionType = get(value).getOrElse(TEXT)
-  }
+  implicit def stringToQuestionType(value: String): QuestionType = get(value).getOrElse(TEXT)
 }
