+++ src/main/scala/com/particeep/api/models/form/edition/QuestionEdition.scala
@@ -2,7 +2,7 @@ package com.particeep.api.models.form.edition
 
 import play.api.libs.json.{ Json, OFormat }
 
-import com.particeep.api.models.enums.QuestionType.QuestionType
+import com.particeep.api.models.enums.QuestionType
 
 case class QuestionEdition(
   label:                 Option[Map[String, String]] = None,
