+++ src/main/scala/com/particeep/api/models/form/get_deep/QuestionDeep.scala
@@ -5,7 +5,7 @@ import play.api.libs.json.{ Json, OFormat, Writes }
 import java.time.OffsetDateTime
 
 import com.particeep.api.core.Formatter
-import com.particeep.api.models.enums.QuestionType.QuestionType
+import com.particeep.api.models.enums.QuestionType
 import com.particeep.api.models.form.get.Possibility
 
 case class QuestionDeep(
